multipleDaeCrash branch - Got the DOM working on Windows again.
git-svn-id: https://collada-dom.svn.sourceforge.net/svnroot/collada-dom/branches/multipleDaeCrash@351 dfea1ed5-6b0d-0410-973e-e49e28f48b26

+++ COLLADA_DOM/test/domTest.cpp
@@ -1243,7 +1243,7 @@ DefineTest(unusedTypeCheck) {
 	const daeMetaElementRefArray &metas = dae.getAllMetas();
 	for (size_t i = 0; i < metas.getCount(); i++)
 		if (!metas[i])
-			actualUnusedTypes.insert(i);
+			actualUnusedTypes.insert((int)i);
 
 	// Make sure the set of unused types matches what we expect
 	return testResult(expectedUnusedTypes == actualUnusedTypes);
